% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkpdmodels.r
\name{alfentMaitre}
\alias{alfentMaitre}
\title{Alfentanil, Maitre 1987 as tested 1988 Anesthesiology}
\usage{
alfentMaitre(age, wt, MALE)
}
\arguments{
\item{age}{Age in years}

\item{wt}{Weight}

\item{MALE}{Gender (MALE=1)}
}
\description{
V1..V3,Cl1..Cl3
}
\examples{
alfentMaitre(20,76,1)
}
\keyword{models}
