% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkpdmodels.r
\name{diprifusor}
\alias{diprifusor}
\title{Propofol, Diprifusor}
\usage{
diprifusor(age, wt)
}
\arguments{
\item{age}{Age in years}

\item{wt}{Weight}
}
\description{
V1..V3,Cl1..Cl3,ke0
}
\examples{
diprifusor(20,76)
}
\keyword{models}
