% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coagulation.r
\name{enoxaparin}
\alias{enoxaparin}
\title{Enoxaparin, Br J Clin Pharmacol 2003; 56, 96–103}
\usage{
enoxaparin(wt, lbm)
}
\arguments{
\item{wt}{Weight}

\item{lbm}{Lean body mass (James)}
}
\description{
Reference time = hour!
}
\examples{
enoxaparin(70,45)
}
\keyword{models}
