% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coagulation.r
\name{enoxaparin2}
\alias{enoxaparin2}
\title{Enoxaparin, Br J Clin Pharmacol. 2007 Oct; 64(4): 428–438.}
\usage{
enoxaparin2(age, wt, male, sCr)
}
\arguments{
\item{age}{Age}

\item{wt}{Weight}

\item{male}{Gender, male?}

\item{sCr}{Serum creatinin}
}
\description{
Reference time = hour!
}
\examples{
enoxaparin2(60,70,1,34)
}
\keyword{models}
