% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coagulation.r
\name{enoxaparin3}
\alias{enoxaparin3}
\title{Enoxaparin,  Br J Clin Pharmacol. 2006 Aug;62(2):165-76.}
\usage{
enoxaparin3(age, wt, IBW, male, sCr)
}
\arguments{
\item{age}{Age}

\item{wt}{Weight}

\item{IBW}{Ideal body weight}

\item{male}{Gender, male?}

\item{sCr}{Serum creatinin}
}
\description{
Reference time = hour!
}
\examples{
enoxaparin3(80,70,63,1,34)
}
\keyword{models}
