% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkpdmodels.r
\name{fentaScott}
\alias{fentaScott}
\title{Scott Fentanyl weight scaled with ke0 0.147}
\usage{
fentaScott(wt)
}
\arguments{
\item{wt}{Weight}
}
\description{
V1..V3,Cl1..Cl3, ke0
}
\examples{
fentaScott(76)
}
\keyword{models}
