% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkpdmodels.r
\name{propEleveld}
\alias{propEleveld}
\title{Propofol, Eleveld BJA 2018}
\usage{
propEleveld(
  wt,
  ht,
  age,
  male,
  pma = 40,
  venous = 0,
  opiates = 1,
  patient = 1,
  tpeak = 0
)
}
\arguments{
\item{wt}{Weight}

\item{ht}{Height}

\item{age}{Age in years}

\item{male}{Gender, male?}

\item{pma}{postmenstrual age}

\item{venous}{Venous concentration?}

\item{opiates}{Are opiates used?}

\item{patient}{Is it a patient?}

\item{tpeak}{Time to peak effect site concentration (1.6 min)}
}
\description{
V1..V3,Cl1..Cl3,ke0
}
\examples{
propEleveld(70,172,40,1)
}
\keyword{models}
