% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkpdmodels.r
\name{propSchnider}
\alias{propSchnider}
\title{Propofol Schnider}
\usage{
propSchnider(age, wt, ht, male)
}
\arguments{
\item{age}{Age in years}

\item{wt}{Weight (kg)}

\item{ht}{Height(cm)}

\item{male}{Gender (male=1)}
}
\description{
V1..V3,Cl1..Cl3,ke0
}
\examples{
propSchnider(20,70,172,1)
}
