% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkpdmodels.r
\name{propSchnider2}
\alias{propSchnider2}
\title{Propofol Schnider (2)}
\usage{
propSchnider2(age, wt, ht, lbm)
}
\arguments{
\item{age}{Age in years}

\item{wt}{Weight (kg)}

\item{ht}{Height(cm)}

\item{lbm}{Lean body mass (James)}
}
\description{
V1..V3,Cl1..Cl3,ke0
}
\examples{
propSchnider(20,70,172,50.6)
}
\keyword{models}
