% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkpdmodels.r
\name{remiMinto}
\alias{remiMinto}
\title{Remifentanil(incl. ke0), Minto et al. (Age, LBM) Anesthesiology 1997; 86:10-23.}
\usage{
remiMinto(age, lbm)
}
\arguments{
\item{age}{Age in years}

\item{lbm}{Lean body mass (James)}
}
\description{
V1..V3,Cl1..Cl3,ke0
}
\examples{
remiMinto(20,50.6)
}
\keyword{models}
